BRAYNS 471 - Adds progressive resolution framebuffer (#1120)

+++ brayns/engine/renderer/Renderer.cpp
@@ -65,7 +65,7 @@ bool Renderer::commit()
     return true;
 }
 
-int32_t Renderer::getSamplesPerPixel() const noexcept
+size_t Renderer::getSamplesPerPixel() const noexcept
 {
     return _data->getSamplesPerPixel();
 }
